[wip] grid functools and struct to grid.h
