--- include/controller.h
@@ -1,7 +1,0 @@
-#ifndef __CONTROLLER_
-#define __CONTROLLER_
-
-void test();
-
-
-#endif
