+++ include/window.h
@@ -1,0 +1,5 @@
+#ifndef __WINDOW_
+#define __WINDOW_
+
+
+#endif
