--- src/controller.c
@@ -1,9 +1,0 @@
-#include"controller.h"
-#include<stdio.h>
-
-void test(){
-
-
-	puts("test controller");
-
-}
